Address thermo-nuclear review items 1–4 on issue remediation PR.
Split stream/bridge tests under 800 lines, break bridge→transcript coupling via record utils, extract CursorSdkTurnCoordinator from streamCursor, and unify transcript/display dispatch in a single tool spec registry.
